backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart
pydantic>=2.0.0
pydantic-settings
azure-storage-blob==12.19.0
openai>=1.0.0
pdfplumber==0.10.4
shapely==2.0.3
azure-ai-formrecognizer==3.3.3
azure-search-documents
azure-ai-documentintelligence
